requirements.txt
transformers
torch
gradio
opencv-python
sounddevice
soundfile
librosa
pandas
datasets
scikit-learn
sentence-transformers
peft
trl
accelerate
bitsandbytes
huggingface_hub
python-dotenv
timm
torchvision
Pillow
numpy
